+++ src/humandesign/features/__init__.py
@@ -10,6 +10,7 @@ from .core import (
     composite_chakras_channels,
     get_composite_combinations,
     get_penta,
+    get_penta_v2,
     hd_composite
 )
 from .attributes import (